requirements.txt
boto3==1.20.37
botocore==1.23.37
jmespath==0.10.0
lxml==4.7.1
numpy==1.22.1
pandas==1.3.5
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
s3transfer==0.5.0
six==1.16.0
urllib3==1.26.8
